Fallback to other modes if websocket is not supported

+++ cloud_cs/session.py
@@ -10,11 +10,10 @@ class Session:
     cookie_name = "cloudcs_sess"
 
     def auth_valid(self, req):
-        self.local_work_dir = None
-        logger.info("authenticated " + self.user_id())
+        logger.info("%s authenticated %s", self.log_str(), str(self.user_id()))
         if not self.cfg().is_user_allowed(self.user_id()):
             raise RuntimeError("user not authorized")
-        logger.debug("user " + self.user_id() + " allowed in role(s) " + str(self.user_role()))
+        logger.debug("%s user %s allowed in role(s) %s", self.log_str(), str(self.user_id()), str(self.user_role()))
         
         self.local_work_dir = None
         # reattach older session if available
@@ -32,17 +31,18 @@ class Session:
         
         if None == self.local_work_dir:
             self.local_work_dir = Utils.mkdtemp(prefix="sess_")
-            logger.debug("created temporary folder " + self.local_work_dir)
+            logger.debug("%s created temporary folder %s", self.log_str(), str(self.local_work_dir))
         else:
-            logger.debug("reattached to old temporary folder " + self.local_work_dir)
+            logger.debug("%s reattached to old temporary folder %s", self.log_str(), str(self.local_work_dir))
         req.set_secure_cookie(Session.cookie_name, self.sess_id, 1)
         
         req.redirect('/auth/storage?uid=' + self.user_id())
             
     def remove_temporary_files(self):
         outdir = self.local_work_dir
-        Utils.rmdir(outdir)
-        logger.debug("removed temporary folder " + outdir)
+        if None != outdir:
+            Utils.rmdir(outdir)
+            logger.debug("%s removed temporary folder %s", self.log_str(), str(outdir))
 
     @staticmethod
     def extract_session_id(req):
@@ -117,6 +117,12 @@ class Session:
     @abstractmethod
     def check_task_timeouts(self):
         raise NotImplementedError
+    
+    def log_str(self):
+        try:
+            return ('[' + str(self.user_id()) + ' - ' + str(self.sess_id) + ']')
+        except:
+            return ''
 
 class SessionStandalone(Session):
     def __init__(self, user):
@@ -126,7 +132,7 @@ class SessionStandalone(Session):
         self.nv = {}
         self.task = None
         self.detach = False
-        logger.debug("created standalone session for " + self.uid + " with key " + self.sess_id)
+        logger.debug("%s session created", self.log_str())
 
     def user_id(self):
         return self.uid
@@ -178,7 +184,7 @@ class SessionInMemory(Session):
         self.sess_id = hashlib.sha1('_'.join([self.uid, self.cfg().cfg_get("SECURE_SALT"), str(self.creation_time)])).hexdigest()
         self.nv = {}
         SessionInMemory.SESS_STORE[self.sess_id] = self
-        logger.debug("created session for " + self.uid + " with key " + self.sess_id)
+        logger.debug("%s session created", self.log_str())
     
     def user_id(self):
         return self.uid
@@ -203,25 +209,26 @@ class SessionInMemory(Session):
     @staticmethod
     def get_session(sess_id):
         sess = None
-        if SessionInMemory._is_valid(sess_id):
-            sess = SessionInMemory.SESS_STORE[sess_id]
-        else:
-            logger.error("session id " + str(sess_id) + " not found")            
+        if None != sess_id:
+            if SessionInMemory._is_valid(sess_id):
+                sess = SessionInMemory.SESS_STORE[sess_id]
+            else:
+                logger.error("session id " + str(sess_id) + " not found")            
         return sess
 
     def logout(self):
-        logger.info("logging out " + self.sess_id)
+        logger.info("%s logging out ", self.log_str())
         self.remove_temporary_files()
         if self.sess_id in SessionInMemory.SESS_STORE.keys():
             del SessionInMemory.SESS_STORE[self.sess_id]
 
     def remove_older_sessions(self):
-        logger.info("getting older sessions of " + self.sess_id + " id:" + self.uid)
+        logger.info("%s getting older sessions...", self.log_str())
         ret = []
         for sess_id in SessionInMemory.SESS_STORE.keys():
             sess = SessionInMemory.SESS_STORE[sess_id]
             if (sess.uid == self.uid) and (sess.sess_id != self.sess_id):
-                logger.info("found old session " + sess.sess_id + " of user id:" + sess.uid)
+                logger.info("%s found old session %s of user id %s", self.log_str(), str(sess.sess_id), str(sess.uid))
                 ret.append(sess)
                 del SessionInMemory.SESS_STORE[sess_id]
         return ret
